packages/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.5.0
openai==1.3.7
pypdf==3.17.1
python-docx==1.1.0
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.3.27
sentence-transformers==5.1.1
psycopg2-binary==2.9.10
transformers>=4.41.0,<5.0.0
torch==2.5.1
numpy>=1.26.0
pydantic-settings==2.0.3
setuptools>=65.5.0
wheel>=0.38.4
